Add full results to MapNotes for location data.
This change adds the fullLocationResults iVar to
MapNote.  This allows a request to have the
full results returned for any non-Kinvey Location
provider.  This is good/bad, since you can now
explore the dictionary to get specific information,
but you can also now break the portability
of the DI classes.